adversarial_experiments/requirements.example.m1.txt
imageio==2.26.1
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==15.0.6.1
matplotlib==3.7.0
numpy==1.24.2
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
pickleshare==0.7.5
Pillow==9.4.0
scikit-image==0.20.0
scikit-learn==1.2.1
scipy==1.10.1
tensorflow-macos==2.10.0
tensorflow-metadata==1.12.0
tensorflow-metal==0.6.0
